// my-app/src/types/typeUsers.ts
type Users = {
    id: string
    name: string
    userName: string
    email: string
    address: {
        street: string
        suite: string
        city: string
        zipcode: string,
        geo: {
            lat: string
            lng: string
        }
    }
    phone: string
    website: string
    company: {
        name: string
        catchPhrase: string
        bs: string
    }
}

export default Users